Fix Planner view and Add restriction DB connection

+++ HomemadeDB/dbo/Tables/Meals.sql
@@ -5,7 +5,7 @@
     [UserId] INT NOT NULL,
     [ExpectedDate] DATE NOT NULL, 
     [NumberOfMeal] INT NOT NULL, 
-    [Servings] DECIMAL NOT NULL,
+    [Servings] INT NOT NULL,
     [Notes] NVARCHAR(50) NULL, 
     CONSTRAINT [FK_Meals_Users] FOREIGN KEY ([UserId]) REFERENCES [Users]([UserId]), 
     CONSTRAINT [FK_Meals_Recepies] FOREIGN KEY ([RecepieId]) REFERENCES [Recepies]([RecepieId])
